refactor:refactor channel list
refactoring channel list improving code readability
