+++ lib/features/presentation/state_management/chat_handler/channel_list_bloc.dart/channel_list_event.dart
@@ -1,0 +1,20 @@
+import 'package:sendbird_chat_sdk/sendbird_chat_sdk.dart';
+
+abstract class ChannelListEvent {
+  const ChannelListEvent();
+}
+
+class RefreshChannels extends ChannelListEvent {
+  final List<GroupChannel> collection;
+  const RefreshChannels(this.collection);
+}
+
+class GetChannels extends ChannelListEvent {
+  final GroupChannelCollectionHandler handler;
+  const GetChannels(this.handler);
+}
+
+class SearchChannels extends ChannelListEvent {
+  final String query;
+  const SearchChannels(this.query);
+}
