+++ lib/features/presentation/state_management/chat_handler/channel_list_bloc.dart/channel_list_state.dart
@@ -1,0 +1,16 @@
+import 'package:sendbird_chat_sdk/sendbird_chat_sdk.dart';
+
+abstract class ChannelListState {
+  final List<GroupChannel>? channelList;
+
+  const ChannelListState({this.channelList});
+}
+
+class ChannelListLoading extends ChannelListState {
+  const ChannelListLoading();
+}
+
+class ChannelListSuccess extends ChannelListState {
+  const ChannelListSuccess(List<GroupChannel> channelList)
+      : super(channelList: channelList);
+}
